+++ CHANGELOG.md
@@ -3,7 +3,9 @@
 
 ## 1.7.0 - 2016-11-11
 
+- Add Command Line Interface (CLI)
 - Add Wikidata provider
+- Add --nearest to Wikidata
 - Seperate tests per providers `test/test.<provider>.ts`
 
 ## 1.6.0 - 2016-11-3